Changed package name from se.yrgo.employee to se.yrgo.employeasy
