+++ employ-easy/src/main/java/se/yrgo/employeasy/exceptions/ConflictException.java
@@ -1,0 +1,9 @@
+package se.yrgo.employeasy.exceptions;
+
+public class ConflictException extends RuntimeException {
+    private static final long serialVersionUID = 2L;
+
+    public ConflictException(String msg) {
+        super(msg);
+    }
+}
