+++ employ-easy/src/main/java/se/yrgo/employeasy/exceptions/ObjectNotFoundException.java
@@ -1,0 +1,10 @@
+package se.yrgo.employeasy.exceptions;
+
+public class ObjectNotFoundException extends RuntimeException {
+
+    private static final long serialVersionUID = 1L;
+
+    public ObjectNotFoundException(String msg) {
+        super(msg);
+    }
+}
