--- employ-easy/src/main/java/se/yrgo/employee/exceptions/ConflictException.java
@@ -1,9 +1,0 @@
-package se.yrgo.employee.exceptions;
-
-public class ConflictException extends RuntimeException {
-    private static final long serialVersionUID = 2L;
-
-    public ConflictException(String msg) {
-        super(msg);
-    }
-}
