--- employ-easy/src/main/java/se/yrgo/employee/exceptions/ObjectNotFoundException.java
@@ -1,10 +1,0 @@
-package se.yrgo.employee.exceptions;
-
-public class ObjectNotFoundException extends RuntimeException {
-
-    private static final long serialVersionUID = 1L;
-
-    public ObjectNotFoundException(String msg) {
-        super(msg);
-    }
-}
