--- employ-easy/src/main/java/se/yrgo/employee/repositories/EmployeeRepository.java
@@ -1,18 +1,0 @@
-package se.yrgo.employee.repositories;
-
-import java.util.List;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-import se.yrgo.employee.entities.Employee;
-
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
-    @Query(value = "SELECT employee.* emp FROM EMPLOYEE WHERE employee.job_title = ?1", nativeQuery = true)
-    List<Employee> findByJobTitle(@Param(value = "jobTitle") String jobTitle);
-
-    @Query(value = "SELECT employee.* emp FROM EMPLOYEE WHERE employee.email = ?1", nativeQuery = true)
-    List<Employee> findByEmail(@Param(value = "email") String email);
-
-    @Query(value = "SELECT * FROM EMPLOYEE e WHERE e.user_id = ?1", nativeQuery = true)
-    Employee findEmployeeByUserId(String userId);
-}
