+++ employ-easy/src/test/java/se/yrgo/employeasy/controllers/LoginSecurityTest.java
@@ -1,0 +1,57 @@
+package se.yrgo.employeasy.controllers;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
+
+import java.nio.charset.StandardCharsets;
+
+import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.user;
+
+@SpringBootTest
+@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
+class LoginSecurityTest {
+
+    @Autowired
+    private WebApplicationContext context;
+
+    private MockMvc mockMvc;
+
+    @BeforeEach
+    public void setup() {
+        mockMvc = MockMvcBuilders
+                .webAppContextSetup(context)
+                .apply(springSecurity())
+                .build();
+    }
+
+    @Test
+    void getAllEmployeesWithoutAccess() throws Exception {
+        this.mockMvc.perform(get("/v1/employees"))
+                .andExpect(status().isUnauthorized());
+    }
+
+    @Test
+    void getAllEmployeesAsAdmin() throws Exception {
+        this.mockMvc.perform(get("/v1/employees")
+                        .with(user("admin").roles("ADMIN")))
+                .andExpect(status().isOk())
+                .andExpect(content().encoding(StandardCharsets.ISO_8859_1));
+    }
+
+    @Test
+    public void greetingReturnsHelloAndUsername() throws Exception {
+        this.mockMvc.perform(get("/v1/greeting").with(user("marten").roles("EMPLOYEE")))
+                .andExpect(status().isOk())
+                .andExpect(content().string("Hello, marten!"));
+    }
+}
